made scouting layout a separate widget
