+++ assets/config/2022config.json
@@ -67,7 +67,7 @@
                                 "required" : true,
                                 "timestamp": false,
                                 "isCommonValue" : false,
-                                "setCommonValue" : false
+                                "setCommonValue" : true
                             },
                             {
                                 "name" : "team #",
@@ -76,7 +76,8 @@
                                 "required" : true,
                                 "timestamp": false,
                                 "isCommonValue" : false,
-                                "setCommonValue" : false
+                                "setCommonValue" : false,
+                                "autofillType" : "singleTeamNumber"
                             },
                             {
                                 "name": "match type",
